// composables/api/role/enums/role.enums.ts
enum RoleNames {
  WEREWOLF = "werewolf",
  BIG_BAD_WOLF = "big-bad-wolf",
  ACCURSED_WOLF_FATHER = "accursed-wolf-father",
  WHITE_WEREWOLF = "white-werewolf",
  VILLAGER = "villager",
  VILLAGER_VILLAGER = "villager-villager",
  SEER = "seer",
  CUPID = "cupid",
  WITCH = "witch",
  HUNTER = "hunter",
  LITTLE_GIRL = "little-girl",
  DEFENDER = "defender",
  ELDER = "elder",
  SCAPEGOAT = "scapegoat",
  IDIOT = "idiot",
  TWO_SISTERS = "two-sisters",
  THREE_BROTHERS = "three-brothers",
  FOX = "fox",
  BEAR_TAMER = "bear-tamer",
  STUTTERING_JUDGE = "stuttering-judge",
  RUSTY_SWORD_KNIGHT = "rusty-sword-knight",
  THIEF = "thief",
  WILD_CHILD = "wild-child",
  WOLF_HOUND = "wolf-hound",
  ANGEL = "angel",
  PIED_PIPER = "pied-piper",
  SCANDALMONGER = "scandalmonger",
  PREJUDICED_MANIPULATOR = "prejudiced-manipulator",
  ACTOR = "actor",
  DEVOTED_SERVANT = "devoted-servant",
}

enum RoleSides {
  VILLAGERS = "villagers",
  WEREWOLVES = "werewolves",
}

enum RoleTypes {
  VILLAGER = "villager",
  WEREWOLF = "werewolf",
  LONELY = "lonely",
  AMBIGUOUS = "ambiguous",
}

enum RoleOrigins {
  CLASSIC = "classic",
  NEW_MOON = "new-moon",
  THE_VILLAGE = "the-village",
  CHARACTERS = "characters",
  UNCONVENTIONAL = "unconventional",
}

export {
  RoleNames,
  RoleSides,
  RoleTypes,
  RoleOrigins,
};